+++ packages/core/binding/binding.d.ts
@@ -41,6 +41,5 @@ export class Compiler {
   update(paths: Array<string>): Promise<JsUpdateResult>;
   /** sync update */
   updateSync(paths: Array<string>): JsUpdateResult;
-  resources(): Record<string, Array<number>>;
   hasModule(resolvedPath: string): boolean;
 }